backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
numpy>=1.26.0
opencv-python>=4.9.0
Pillow>=10.2.0
scipy>=1.12.0
scikit-image>=0.22.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
